+++ pkg/providers/property/testdata/TestResourceEdgeHostname/new_akamaized_update_ttl.tf
@@ -1,0 +1,17 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_edge_hostname" "edgehostname" {
+  contract_id         = "ctr_2"
+  group_id            = "grp_2"
+  product_id          = "prd_2"
+  edge_hostname       = "test.akamaized.net"
+  ip_behavior         = "IPV6_COMPLIANCE"
+  status_update_email = ["hello@akamai.com"]
+  ttl                 = 800
+}
+
+output "edge_hostname" {
+  value = akamai_edge_hostname.edgehostname.edge_hostname
+}